+++ NEWS.md
@@ -1,3 +1,12 @@
+# fabricQueryR 0.2.1
+
+* Update e-mail address of maintainer in DESCRIPTION file (change to a 
+personal e-mail address due to leaving the organization).
+
+* `fabric_onelake_read_delta_table()`: add experimental support for specifying
+a Lakehouse schema name to read from a specific schema within a Lakehouse which
+has Lakehouse schemas enabled.
+
 # fabricQueryR 0.2.0
 
 * Added `fabric_livy_query()` to run queries against the 'Fabric Livy API',